refactor(unify): Phase 1 prep - relocate shared decls to break package cycles
Pure relocations inside the flat core package, no behavior change, so the
upcoming package split has an acyclic file graph to work with:

- extractStringArray/extractFeedURLsFromConfig, setCORSHeaders -> helpers.go
- InvalidateOverviewCache + RedisOverviewCachePrefix -> redis.go
- recent-ticket window (Push/Fetch + keys) -> ai_triage.go, beside its type
- tierFromRoles -> users.go
- publicFeedGroup -> handlers_public.go
- GetUserWidgets -> models.go, beside the Widget model

Also:
- nil-guard UpdateUserTopicSubscriptions so widget writes never panic when
  the SSE hub was never initialized (lets tests drop hand-built hubs)
- ResetM2MTokenCache replaces three exported-for-test mutable globals
