+++ api/core/redis.go
@@ -2,7 +2,6 @@ package core
 
 import (
 	"context"
-	"encoding/json"
 	"log"
 	"os"
 
@@ -116,51 +115,26 @@ func InvalidateUserCaches(userSub string) {
 	}
 }
 
-// --- AI Triage: Recent Ticket Summaries ---
-// Sliding window of the last N ticket summaries, used as context when
-// asking Claude to dupe-detect against recent submissions. Keyed by a
-// single global list; entries are LPUSHed and trimmed to the cap.
+// RedisOverviewCachePrefix is the per-user key prefix for the overview
+// cache. Format: overview:{logto_sub}.
+const RedisOverviewCachePrefix = "overview:"
 
-const (
-	RedisRecentTicketsKey   = "support:recent_tickets"
-	RedisRecentTicketsLimit = 50
-)
-
-// PushRecentTicketSummary adds a summary to the head of the sliding
-// window. Atomic — uses LPUSH+LTRIM in a pipeline.
-func PushRecentTicketSummary(ctx context.Context, summary RecentTicketSummary) {
-	if Rdb == nil {
+// InvalidateOverviewCache deletes the per-user overview cache key.
+// Called from the Stripe webhook (subscription state changes), the
+// widget CRUD handlers (toggle state changes), and the GDPR request
+// lifecycle (deletion status changes) so the next request always sees
+// fresh data instead of waiting up to OverviewCacheTTL for the cache
+// to expire.
+//
+// Failures are logged and swallowed — the caller's primary write has
+// already succeeded; an invalidation miss only delays the visible
+// effect by OverviewCacheTTL, which is acceptable.
+func InvalidateOverviewCache(ctx context.Context, userID string) {
+	if userID == "" || Rdb == nil {
 		return
 	}
-	bytes, err := json.Marshal(summary)
-	if err != nil {
-		log.Printf("[Redis] marshal recent ticket: %v", err)
-		return
+	key := RedisOverviewCachePrefix + userID
+	if err := Rdb.Del(ctx, key).Err(); err != nil {
+		log.Printf("[Overview] cache invalidate failed for %s: %v", userID, err)
 	}
-	pipe := Rdb.Pipeline()
-	pipe.LPush(ctx, RedisRecentTicketsKey, string(bytes))
-	pipe.LTrim(ctx, RedisRecentTicketsKey, 0, int64(RedisRecentTicketsLimit-1))
-	if _, err := pipe.Exec(ctx); err != nil {
-		log.Printf("[Redis] push recent ticket: %v", err)
-	}
-}
-
-// FetchRecentTicketSummaries returns the last N summaries (most recent first).
-func FetchRecentTicketSummaries(ctx context.Context) []RecentTicketSummary {
-	if Rdb == nil {
-		return nil
-	}
-	rawList, err := Rdb.LRange(ctx, RedisRecentTicketsKey, 0, int64(RedisRecentTicketsLimit-1)).Result()
-	if err != nil {
-		log.Printf("[Redis] fetch recent tickets: %v", err)
-		return nil
-	}
-	out := make([]RecentTicketSummary, 0, len(rawList))
-	for _, s := range rawList {
-		var sum RecentTicketSummary
-		if err := json.Unmarshal([]byte(s), &sum); err == nil {
-			out = append(out, sum)
-		}
-	}
-	return out
 }
